// src/gql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Event = {
  __typename?: 'Event';
  /** How long before the event start will the check-in end (in seconds) */
  checkInBuffer?: Maybe<Scalars['Int']['output']>;
  /** How long the event check-in will last (in seconds) */
  checkInDuration?: Maybe<Scalars['Int']['output']>;
  /** Whether check-in is enabled for this event */
  checkInEnabled?: Maybe<Scalars['Boolean']['output']>;
  /** Rough categorization of event tier, denoting relative importance in the competitive scene */
  competitionTier?: Maybe<Scalars['Int']['output']>;
  /** When the event was created (unix timestamp) */
  createdAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Last date attendees are able to create teams for team events */
  deckSubmissionDeadline?: Maybe<Scalars['Timestamp']['output']>;
  /**
   * Maximum number of participants each Entrant can have
   * @deprecated Migrate to teamRosterSize
   */
  entrantSizeMax?: Maybe<Scalars['Int']['output']>;
  /**
   * Minimum number of participants each Entrant can have
   * @deprecated Migrate to teamRosterSize
   */
  entrantSizeMin?: Maybe<Scalars['Int']['output']>;
  /** The entrants that belong to an event, paginated by filter criteria */
  entrants?: Maybe<EntrantConnection>;
  /** How much it costs to enter this event */
  entryFee?: Maybe<Scalars['Float']['output']>;
  /** Whether the event has decks */
  hasDecks?: Maybe<Scalars['Boolean']['output']>;
  /** Are player tasks enabled for this event */
  hasTasks?: Maybe<Scalars['Boolean']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  /** Whether the event is an online event or not */
  isOnline?: Maybe<Scalars['Boolean']['output']>;
  league?: Maybe<League>;
  /** Markdown field for match rules/instructions */
  matchRulesMarkdown?: Maybe<Scalars['String']['output']>;
  /** Title of event set by organizer */
  name?: Maybe<Scalars['String']['output']>;
  /** Gets the number of entrants in this event */
  numEntrants?: Maybe<Scalars['Int']['output']>;
  /** The phase groups that belong to an event. */
  phaseGroups?: Maybe<Array<Maybe<PhaseGroup>>>;
  /** The phases that belong to an event. */
  phases?: Maybe<Array<Maybe<Phase>>>;
  /** TO settings for prizing */
  prizingInfo?: Maybe<Scalars['JSON']['output']>;
  publishing?: Maybe<Scalars['JSON']['output']>;
  /** Markdown field for event rules/instructions */
  rulesMarkdown?: Maybe<Scalars['String']['output']>;
  /** Id of the event ruleset */
  rulesetId?: Maybe<Scalars['Int']['output']>;
  /**
   * Settings pulled from the event ruleset, if one exists
   * @deprecated Use ruleset
   */
  rulesetSettings?: Maybe<Scalars['JSON']['output']>;
  /** Paginated sets for this Event */
  sets?: Maybe<SetConnection>;
  slug?: Maybe<Scalars['String']['output']>;
  /** Paginated list of standings */
  standings?: Maybe<StandingConnection>;
  /** When does this event start? */
  startAt?: Maybe<Scalars['Timestamp']['output']>;
  /** The state of the Event. */
  state?: Maybe<ActivityState>;
  /** Paginated stations on this event */
  stations?: Maybe<StationsConnection>;
  /** Last date attendees are able to create teams for team events */
  teamManagementDeadline?: Maybe<Scalars['Timestamp']['output']>;
  /** If this is a teams event, returns whether or not teams can set custom names */
  teamNameAllowed?: Maybe<Scalars['Boolean']['output']>;
  /** Team roster size requirements */
  teamRosterSize?: Maybe<TeamRosterSize>;
  tournament?: Maybe<Tournament>;
  /** The type of the event, whether an entrant will have one participant or multiple */
  type?: Maybe<Scalars['Int']['output']>;
  /** When the event was last modified (unix timestamp) */
  updatedAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Whether the event uses the new EventSeeds for seeding */
  useEventSeeds?: Maybe<Scalars['Boolean']['output']>;
  /** The entrant (if applicable) for a given user in this event */
  userEntrant?: Maybe<Entrant>;
  videogame?: Maybe<Videogame>;
  /** The waves being used by the event */
  waves?: Maybe<Array<Maybe<Wave>>>;
};
// ...
export type Game = {
  __typename?: 'Game';
  /** Score of entrant 1. For smash, this is equivalent to stocks remaining. */
  entrant1Score?: Maybe<Scalars['Int']['output']>;
  /** Score of entrant 2. For smash, this is equivalent to stocks remaining. */
  entrant2Score?: Maybe<Scalars['Int']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  orderNum?: Maybe<Scalars['Int']['output']>;
  /** Selections for this game such as character, etc. */
  selections?: Maybe<Array<Maybe<GameSelection>>>;
  /** The stage that this game was played on (if applicable) */
  stage?: Maybe<Stage>;
  state?: Maybe<Scalars['Int']['output']>;
  winnerId?: Maybe<Scalars['Int']['output']>;
};
// ...
export enum GameSelectionType {
  /** Character selection */
  Character = 'CHARACTER'
}
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  filter?: Maybe<Scalars['JSON']['output']>;
  page?: Maybe<Scalars['Int']['output']>;
  perPage?: Maybe<Scalars['Int']['output']>;
  sortBy?: Maybe<Scalars['String']['output']>;
  total?: Maybe<Scalars['Int']['output']>;
  totalPages?: Maybe<Scalars['Int']['output']>;
};
// ...
export type Player = {
  __typename?: 'Player';
  gamerTag?: Maybe<Scalars['String']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  prefix?: Maybe<Scalars['String']['output']>;
  /** Most recent active & published rankings */
  rankings?: Maybe<Array<Maybe<PlayerRank>>>;
  /**
   * Recent sets for this player.
   * @deprecated Use the sets field instead.
   */
  recentSets?: Maybe<Array<Maybe<Set>>>;
  /** Recent standings */
  recentStandings?: Maybe<Array<Maybe<Standing>>>;
  /** Set history for this player. */
  sets?: Maybe<SetConnection>;
  user?: Maybe<User>;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Returns the authenticated user */
  currentUser?: Maybe<User>;
  /** Returns an entrant given its id */
  entrant?: Maybe<Entrant>;
  /** Returns an event given its id or slug */
  event?: Maybe<Event>;
  /** Returns a league given its id or slug */
  league?: Maybe<League>;
  /** Paginated, filterable list of leagues */
  leagues?: Maybe<LeagueConnection>;
  /** Returns a participant given its id */
  participant?: Maybe<Participant>;
  /** Returns a phase given its id */
  phase?: Maybe<Phase>;
  /** Returns a phase group given its id */
  phaseGroup?: Maybe<PhaseGroup>;
  /** Returns a player given an id */
  player?: Maybe<Player>;
  /** Returns a phase seed given its id */
  seed?: Maybe<Seed>;
  /** Returns a set given its id */
  set?: Maybe<Set>;
  /** A shop entity */
  shop?: Maybe<Shop>;
  /** Returns an stream given its id */
  stream?: Maybe<Streams>;
  /** Returns all the stream queues for a given tournament */
  streamQueue?: Maybe<Array<Maybe<StreamQueue>>>;
  /** Returns a team given its id */
  team?: Maybe<Team>;
  /** Returns a tournament given its id or slug */
  tournament?: Maybe<Tournament>;
  /** Paginated, filterable list of tournaments */
  tournaments?: Maybe<TournamentConnection>;
  /** Returns a user given a user slug of the form user/abc123, or id */
  user?: Maybe<User>;
  /** Returns a videogame given its id */
  videogame?: Maybe<Videogame>;
  /** Returns paginated list of videogames matching the search criteria. */
  videogames?: Maybe<VideogameConnection>;
};
// ...
export type Set = {
  __typename?: 'Set';
  /** The time this set was marked as completed */
  completedAt?: Maybe<Scalars['Timestamp']['output']>;
  /** The time this set was created */
  createdAt?: Maybe<Scalars['Timestamp']['output']>;
  displayScore?: Maybe<Scalars['String']['output']>;
  /** The source of the first entrant in this set */
  entrant1Source?: Maybe<SetEntrantSource>;
  /** The source of the second entrant in this set */
  entrant2Source?: Maybe<SetEntrantSource>;
  /** Event that this set belongs to. */
  event?: Maybe<Event>;
  /** Full round text of this set. */
  fullRoundText?: Maybe<Scalars['String']['output']>;
  game?: Maybe<Game>;
  games?: Maybe<Array<Maybe<Game>>>;
  /** Whether this set contains a placeholder entrant */
  hasPlaceholder?: Maybe<Scalars['Boolean']['output']>;
  id?: Maybe<Scalars['ID']['output']>;
  /** The letters that describe a unique identifier within the pool. Eg. F, AT */
  identifier?: Maybe<Scalars['String']['output']>;
  images?: Maybe<Array<Maybe<Image>>>;
  lPlacement?: Maybe<Scalars['Int']['output']>;
  /** The progression seed that the loser of this set will be placed into (if applicable) */
  loserProgressionSeed?: Maybe<Seed>;
  /** Phase group that this Set belongs to. */
  phaseGroup?: Maybe<PhaseGroup>;
  /** The sets that are affected from resetting this set */
  resetAffectedData?: Maybe<ResetAffectedData>;
  /** The round number of the set. Negative numbers are losers bracket */
  round?: Maybe<Scalars['Int']['output']>;
  /**
   * Indicates whether the set is in best of or total games mode. This instructs
   * which field is used to figure out how many games are in this set.
   */
  setGamesType?: Maybe<Scalars['Int']['output']>;
  /** A possible spot in a set. Use this to get all entrants in a set. Use this for all bracket types (FFA, elimination, etc) */
  slots?: Maybe<Array<Maybe<SetSlot>>>;
  /** The start time of the Set. If there is no startAt time on the Set, will pull it from phaseGroup rounds configuration. */
  startAt?: Maybe<Scalars['Timestamp']['output']>;
  startedAt?: Maybe<Scalars['Timestamp']['output']>;
  state?: Maybe<Scalars['Int']['output']>;
  /** Tournament event station for a set */
  station?: Maybe<Stations>;
  /** Tournament event stream for a set */
  stream?: Maybe<Streams>;
  /** If setGamesType is in total games mode, this defined the number of games in the set. */
  totalGames?: Maybe<Scalars['Int']['output']>;
  /** The time this set was last updated */
  updatedAt?: Maybe<Scalars['Timestamp']['output']>;
  /** Url of a VOD for this set */
  vodUrl?: Maybe<Scalars['String']['output']>;
  wPlacement?: Maybe<Scalars['Int']['output']>;
  winnerId?: Maybe<Scalars['Int']['output']>;
  /** The progression seed that the winner of this set will be placed into (if applicable) */
  winnerProgressionSeed?: Maybe<Seed>;
};


/** A set */
export type SetDisplayScoreArgs = {
  mainEntrantId?: InputMaybe<Scalars['ID']['input']>;
};


/** A set */
export type SetGameArgs = {
  orderNum: Scalars['Int']['input'];
};


/** A set */
export type SetImagesArgs = {
  type?: InputMaybe<Scalars['String']['input']>;
};


/** A set */
export type SetSlotsArgs = {
  includeByes?: InputMaybe<Scalars['Boolean']['input']>;
};

export type SetConnection = {
  __typename?: 'SetConnection';
  nodes?: Maybe<Array<Maybe<Set>>>;
  pageInfo?: Maybe<PageInfo>;
};
// ...
export type GetPlayerSetsQueryVariables = Exact<{
  playerId: Scalars['ID']['input'];
  page: Scalars['Int']['input'];
  perPage: Scalars['Int']['input'];
}>;


export type GetPlayerSetsQuery = { __typename?: 'Query', player?: { __typename?: 'Player', id?: string | null, sets?: { __typename?: 'SetConnection', pageInfo?: { __typename?: 'PageInfo', totalPages?: number | null } | null, nodes?: Array<{ __typename?: 'Set', id?: string | null, fullRoundText?: string | null, displayScore?: string | null, totalGames?: number | null, winnerId?: number | null, completedAt?: any | null, event?: { __typename?: 'Event', id?: string | null, name?: string | null, tournament?: { __typename?: 'Tournament', id?: string | null, name?: string | null } | null } | null, games?: Array<{ __typename?: 'Game', id?: string | null, selections?: Array<{ __typename?: 'GameSelection', selectionType?: GameSelectionType | null, selectionValue?: number | null, entrant?: { __typename?: 'Entrant', id?: string | null } | null } | null> | null } | null> | null, slots?: Array<{ __typename?: 'SetSlot', entrant?: { __typename?: 'Entrant', id?: string | null, name?: string | null, participants?: Array<{ __typename?: 'Participant', id?: string | null, player?: { __typename?: 'Player', id?: string | null } | null, user?: { __typename?: 'User', id?: string | null } | null } | null> | null } | null } | null> | null } | null> | null } | null } | null };